Summary of what i added
I made changes to the receipt manager so that the background can work. To do that, you need to copy the receipt manager in dev2 scene, and also the gameobject named "levelBackgrounds", you might have to drag the sprites onto the receipt manager, leaving element 0 blank because that does not represent any of the levels (put limbo in element 1). Also, if you don't have it, you need to drag the UI manager into dev 1 too, and also the "day" and "trust" gameobject. Drag them onto the corresponding spot on the ui manager.